refactor: #390 JWT token 관련 business logic을 CQRS pattern 적용하여 분리
