+++ src/main/java/com/zelusik/eatery/global/auth/service/JwtTokenCommandService.java
@@ -6,16 +6,16 @@ import com.zelusik.eatery.global.auth.entity.RefreshToken;
 import com.zelusik.eatery.global.auth.exception.RefreshTokenValidateException;
 import com.zelusik.eatery.global.auth.exception.TokenValidateException;
 import com.zelusik.eatery.global.auth.repository.RefreshTokenRepository;
-import com.zelusik.eatery.global.security.dto.JwtTokenInfoDto;
 import com.zelusik.eatery.global.security.JwtTokenProvider;
+import com.zelusik.eatery.global.security.dto.JwtTokenInfoDto;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 @RequiredArgsConstructor
-@Transactional(readOnly = true)
+@Transactional
 @Service
-public class JwtTokenService {
+public class JwtTokenCommandService {
 
     private final RefreshTokenRepository refreshTokenRepository;
     private final JwtTokenProvider jwtTokenProvider;
@@ -29,7 +29,6 @@ public class JwtTokenService {
      * @param loginType jwt token을 생성하고자 하는 회원의 login type
      * @return 생성된 access token과 refresh token 정보가 담긴 <code>TokenResponse</code> 객체
      */
-    @Transactional
     public JwtTokenDto create(Long memberId, LoginType loginType) {
         JwtTokenInfoDto accessTokenInfo = jwtTokenProvider.createAccessToken(memberId, loginType);
         JwtTokenInfoDto refreshTokenInfo = jwtTokenProvider.createRefreshToken(memberId, loginType);
@@ -50,7 +49,6 @@ public class JwtTokenService {
      * @return 새롭게 생성된 access token과 refresh token 정보가 담긴 <code>TokenResponse</code> 객체
      * @throws TokenValidateException 유효하지 않은 token인 경우
      */
-    @Transactional
     public JwtTokenDto refresh(String oldRefreshToken) {
         jwtTokenProvider.validateToken(oldRefreshToken);
 
@@ -61,23 +59,5 @@ public class JwtTokenService {
                 oldRedisRefreshToken.getMemberId(),
                 jwtTokenProvider.getLoginType(oldRefreshToken)
         );
-    }
-
-    /**
-     * <p>
-     * Refresh token의 유효성을 검사한다.
-     * <p>
-     * Refresh token이 유효하지 않은 값인 경우, refresh token이 만료된 경우가 유효하지 않은 경우이다.
-     *
-     * @param refreshToken 유효성을 검사할 refreshToken
-     * @return refresh token의 유효성 검사 결과
-     */
-    public boolean validateOfRefreshToken(String refreshToken) {
-        try {
-            jwtTokenProvider.validateToken(refreshToken);
-        } catch (Exception ex) {
-            return false;
-        }
-        return refreshTokenRepository.existsById(refreshToken);
     }
 }
